back/requirements.txt
django
djangorestframework
markdown
django-filter
tqdm
django-cors-headers